Be less strict about visibility for warnings on synthetic elements
Fixes compat with timezone 0.8.0

We try to place error messages for synthetic elements at
either its setter or at its getter. Before now, that
setter/getter had to be both public and synthetic; now,
we'll just print a warning if that check fails. Note that
we still error out if we can't find any setter or getter.

Related to #2034