Derive Parameters for unusual function aliases
Fixes compat with petitparser 4.4.0

If a TypeAliasElement aliases to a FunctionType, the
`aliasedElement` property might be null, and our code
would then mark it as not callable. This change makes
`ModelElement.parameters` understand how to retrieve info
out of a FunctionType and thus let us treat the element as
callable.